Inheritance
This is a collection of academic C++ codes where we visualize the idea of Inheritance:

"Inheritance is a mechanism in C++ by which one class can acquire the properties (data members) and behaviors (member functions) of another class. It supports the concept of hierarchical classification. Inheritance provides an opportunity to reuse code and also allows adding additional features to an existing class without modifying it."

Source:
URL: https://www.tutorialspoint.com/cplusplus/cpp_inheritance.html